Add proxy support on build script
We now have a variable called `PROXY`, which will be (I think you
guessed...) the proxy address used by the `docker build` command.

Currently, the proxy address is for both HTTP and HTTPS protocols.
Perhaps we should make a distinction in the future.

I also bumped the version of Microsoft.Spark library to the latest
avaiable, which is 0.12.1